# Faster and simpler sprand for SparseMatrixCSC