The code for sprand for `SparseMatrixCSC` picks nonempty columns first, and then fill them (with the conditional probability of being nonempty). This seems to be done for faster construction for very sparse matrices (so that empty columns are to be expected). I think however that this does not give a real advantage, because one still needs to fill the colptr array, even for empty columns. I implemented a simpler version, and seems to be twice as fast or more in most cases. Am I overlooking something? Some benchmarks:
```
for n=(10,100,1000), p=(1.0,1e-1,1e-2,1e-3,1e-4)
    @show(n,p)
    @show @benchmark sprand($n,$n,$p)
end
```

MASTER:

-------------------------------------------------------------
 |    n\p        | 1.0         |   0.1     | 0.01          |  0.001       | 0.0001
|-----------|------------|-----------|---------------|---------------|-------------
| n=10     | 2.458 μs | 2.135 μs | 830.200 ns | 704.667 ns | 688.295 ns
| n=100   | 82.119 μs | 56.133 μs | 13.024 μs | 2.480 μs | 895.000 ns
| n=1000 | 9.556 ms | 3.858 ms | 555.889 μs | 152.445 μs | 17.662 μs
----------------------------------------------------------------------------------

PR:

-----------------------------------------------------------------------------
|   n\p     | 1.0         |   0.1     | 0.01          |  0.001       | 0.0001
|-----------|------------|-----------|---------------|---------------|-------------
| n=10     |751.574 ns|835.169 ns|419.209 ns|355.906 ns|348.586 ns
| n=100   |45.407 μs|32.657 μs|4.680 μs|970.100 ns|586.613 ns
| n=1000 |4.732 ms|3.362 ms|340.277 μs|45.028 μs|7.101 μs
---------------------------------------------------------------------------------

[Raw data](https://github.com/JuliaLang/julia/files/2705628/bench3.txt)

EDIT: Some simple innest-loop optimization improves speed further:

-----------------------------------------------------------------------------
|   n\p     | 1.0         |   0.1     | 0.01          |  0.001       | 0.0001
|-----------|------------|-----------|---------------|---------------|-------------
| n=10     |  361.784 ns|    414.073 ns|   210.166 ns|  170.565 ns|  168.729 ns
| n=100   |24.113 μs|  17.407 μs|    2.821 μs|   700.679 ns|  405.546 ns
| n=1000 |3.373 ms | 1.223 ms | 134.972 μs|  24.485 μs|  5.988 μs
---------------------------------------------------------------------------------