@haampie 2018-12-23
Would this be considered a breaking change as the same random seed will now produce different sparse matrices?