@ViralBShah 2018-12-23
I don't understand (have to look closely) what impact this has on the produced distribution.

It is a breaking change, but it will be in stdlib.