@stevengj 2018-12-24
At first glance, it seems like this will produce the same distribution and is indeed a lot simpler.