@stevengj 2018-12-24
(The original algorithm is from #6726; see also the discussion in #6708.  My fault for not noticing that one could simply use `randsubseq(1:m*n)`; I was too focused on the CSC format as a collection of columns.)