@abraunst 2018-12-24
> I don't understand (have to look closely) what impact this has on the produced distribution.
> 
> It is a breaking change, but it will be in stdlib.

Should be the same distribution (each of the n^2 entries is nonzero independently with probability p). It outputs a different random matrix because it uses fewer random draws. In addition, it allocates only the output arrays (it needs no temp arrays).