@tkf 2018-12-25
What I agree is that the performance improvement is awesome and we should have it (thanks for the PR!).  What I'm not sure is to allow "non-core" random functions to change with less consideration.  If Random.jl come up with some compatibility policy on versioning, I think it should be uniformly applied to any post-1.0 Julia packages defining random functions.  In any case, I think an ideal situation would be that such libraries are completely decoupled from Julia's release cycle (though I know it's hard for Random.jl) so that "freezing" the random number algorithms per project can easily be done by a Project.toml.  We can't do such decoupling at the moment so I guess I'll trust core devs decisions.