@StefanKarpinski 2019-01-02
cc @rfourquet, @KristofferC