@abraunst 2019-01-03
I've added some tests making output distribution of `sprand` crystal clear.