@ViralBShah 2019-01-04
I don't think we ever made a promise around the structure of sparse random matrices. I think it is fine to merge.