@abraunst 2019-01-05
> I don't think we ever made a promise around the structure of sparse random matrices. I think it is fine to merge.

We could add a pointer to the old version of the method in the news file so people needing to reproduce the old matrices could just copy-paste the old code. The fact that the library is written in julia itself makes this operation particularly painless (wrt. e.g. python) ...