@KristofferC 2019-01-05
I don't think we can backport this to 1.1.1 since it changes the random number stream. Tentatively removing the backport label.