@StefanKarpinski 2019-01-05
We absolutely cannot change this in a point release. Even changing it in a minor release is questionable.