@ViralBShah 2019-01-06
Ah. Yes 1.2 is just fine.