@jarlebring 2019-07-09
From discussion with @fredrikekre in slack: I want to point out (which is suggested above and already discussed in https://github.com/oxfordcontrol/COSMO.jl/issues/83) that this PR leads to different matrices. 

```julia
julia> versioninfo()
Julia Version 1.2.0-rc2.0
julia> using Random, SparseArrays
julia> Random.seed!(0); A=sprandn(3,3,0.2)
3×3 SparseMatrixCSC{Float64,Int64} with 4 stored entries:
  [3, 1]  =  1.57433
  [1, 2]  =  -0.688907
  [1, 3]  =  -0.762804
  [2, 3]  =  0.397482
```


```julia
​julia> versioninfo()
Julia Version 1.1.0
julia> using Random, SparseArrays
julia> Random.seed!(0); A=sprandn(3,3,0.2)
3×3 SparseMatrixCSC{Float64,Int64} with 1 stored entry:
  [1, 3]  =  0.586617
```

The proposed solution of oxfordcontrol to hard-code the matrices is not optimal for us since we want to have a dynamic size.  Our use-case is that we want to present randomly generated sparse matrices as examples in research papers, so that others easily can use the same examples. If the output of `sprandn` and other random generators are not considered stable, this becomes more complicated, i.e., we have to provide a large number of files (which needs to be loaded requiring MAT.jl or similar csvfiles which is not compressed). Compare this to the sprandn-oneliner.